src/api/requirements.txt
fastapi==0.109.2
python-multipart==0.0.9
pillow==10.2.0
ultralytics==8.1.27
tensorflow==2.15.0
uvicorn==0.27.1
numpy==1.26.4